requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
pdf2image
pytesseract
pdfplumber
google-generativeai